+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## 12.x.x (unreleased)
 
 ### Improvements
+- Client: Preferences Registry
 
 ### Bugfixes
 -